+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased (????-??-??)
 
-...
+* fix(stream): consider `ENOTFOUND` (macOS DNS error) as retryable for ApplicationLogStream
+* feat(stream): make `retryCount` public on ApplicationLogStream
+* fix(stream): prevent uncaught errors on retry with ApplicationLogStream (firefox)
+* fix(stream): consider browser network errors as retryable
+* feat(stream): add a `connectionTimeout` param on ApplicationLogStream (default: 5s)
 
 ## 8.0.3 (2024-01-30)